Move FindAndReadDeclarationVariableNode class out from FindDeclarationVariableNodes
